pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "detectkit"
version = "0.1.0"
description = "Metric monitoring with automatic anomaly detection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "detectkit team"}
]
keywords = ["monitoring", "anomaly-detection", "metrics", "timeseries", "alerting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "jinja2>=3.0",
    "orjson>=3.0",
]

[project.optional-dependencies]
# Database drivers
clickhouse = ["clickhouse-driver>=0.2.0"]
postgres = ["psycopg2-binary>=2.9.0"]
mysql = ["pymysql>=1.0.0"]
all-db = ["clickhouse-driver>=0.2.0", "psycopg2-binary>=2.9.0", "pymysql>=1.0.0"]

# Advanced detectors
prophet = ["prophet>=1.1.0"]
timesfm = ["timesfm>=0.1.0"]
advanced-detectors = ["prophet>=1.1.0", "timesfm>=0.1.0"]

# All extras
all = [
    "clickhouse-driver>=0.2.0",
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.0.0",
    "prophet>=1.1.0",
    "timesfm>=0.1.0",
]

# Development dependencies
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
dtk = "detectkit.cli.main:cli"

[project.urls]
Homepage = "https://github.com/alexeiveselov92/detectkit"
Documentation = "https://github.com/alexeiveselov92/detectkit"
Repository = "https://github.com/alexeiveselov92/detectkit"
Issues = "https://github.com/alexeiveselov92/detectkit/issues"

[tool.setuptools]
packages = ["detectkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--cov=detectkitit",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "clickhouse_driver.*",
    "orjson.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
